created waveform

+++ function.h
@@ -4,6 +4,6 @@
 
 void ulcd_display(int i);
 
-void wave(int j);
+void wave(float j);
 
 void sampling();
